backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
google-generativeai>=0.8.0
httpx==0.27.2
python-multipart==0.0.12
aiohttp==3.10.10
requests==2.32.3
python-dateutil==2.9.0
motor==3.6.0
pymongo==4.10.1
amadeus==12.0.0
